Split example components by component type
