+++ component/componenttest/example_receiver.go
@@ -1,0 +1,158 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package componenttest
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/config/configerror"
+	"go.opentelemetry.io/collector/config/configmodels"
+	"go.opentelemetry.io/collector/config/confignet"
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// ExampleReceiver is for testing purposes. We are defining an example config and factory
+// for "examplereceiver" receiver type.
+type ExampleReceiver struct {
+	configmodels.ReceiverSettings `mapstructure:",squash"` // squash ensures fields are correctly decoded in embedded struct
+	// Configures the receiver server protocol.
+	confignet.TCPAddr `mapstructure:",squash"` // squash ensures fields are correctly decoded in embedded struct
+
+	ExtraSetting     string            `mapstructure:"extra"`
+	ExtraMapSetting  map[string]string `mapstructure:"extra_map"`
+	ExtraListSetting []string          `mapstructure:"extra_list"`
+
+	// FailTraceCreation causes CreateTracesReceiver to fail. Useful for testing.
+	FailTraceCreation bool `mapstructure:"-"`
+
+	// FailMetricsCreation causes CreateMetricsReceiver to fail. Useful for testing.
+	FailMetricsCreation bool `mapstructure:"-"`
+}
+
+// ExampleReceiverFactory is factory for ExampleReceiver.
+type ExampleReceiverFactory struct {
+}
+
+var _ component.ReceiverFactory = (*ExampleReceiverFactory)(nil)
+
+// Type gets the type of the Receiver config created by this factory.
+func (f *ExampleReceiverFactory) Type() configmodels.Type {
+	return "examplereceiver"
+}
+
+// CreateDefaultConfig creates the default configuration for the Receiver.
+func (f *ExampleReceiverFactory) CreateDefaultConfig() configmodels.Receiver {
+	return &ExampleReceiver{
+		ReceiverSettings: configmodels.ReceiverSettings{
+			TypeVal: f.Type(),
+			NameVal: string(f.Type()),
+		},
+		TCPAddr: confignet.TCPAddr{
+			Endpoint: "localhost:1000",
+		},
+		ExtraSetting:     "some string",
+		ExtraMapSetting:  nil,
+		ExtraListSetting: nil,
+	}
+}
+
+// CreateTraceReceiver creates a trace receiver based on this config.
+func (f *ExampleReceiverFactory) CreateTracesReceiver(
+	_ context.Context,
+	_ component.ReceiverCreateParams,
+	cfg configmodels.Receiver,
+	nextConsumer consumer.TracesConsumer,
+) (component.TracesReceiver, error) {
+	if cfg.(*ExampleReceiver).FailTraceCreation {
+		return nil, configerror.ErrDataTypeIsNotSupported
+	}
+
+	receiver := f.createReceiver(cfg)
+	receiver.TraceConsumer = nextConsumer
+
+	return receiver, nil
+}
+
+func (f *ExampleReceiverFactory) createReceiver(cfg configmodels.Receiver) *ExampleReceiverProducer {
+	// There must be one receiver for all data types. We maintain a map of
+	// receivers per config.
+
+	// Check to see if there is already a receiver for this config.
+	receiver, ok := exampleReceivers[cfg]
+	if !ok {
+		receiver = &ExampleReceiverProducer{}
+		// Remember the receiver in the map
+		exampleReceivers[cfg] = receiver
+	}
+
+	return receiver
+}
+
+// CreateMetricsReceiver creates a metrics receiver based on this config.
+func (f *ExampleReceiverFactory) CreateMetricsReceiver(
+	_ context.Context,
+	_ component.ReceiverCreateParams,
+	cfg configmodels.Receiver,
+	nextConsumer consumer.MetricsConsumer,
+) (component.MetricsReceiver, error) {
+	if cfg.(*ExampleReceiver).FailMetricsCreation {
+		return nil, configerror.ErrDataTypeIsNotSupported
+	}
+
+	receiver := f.createReceiver(cfg)
+	receiver.MetricsConsumer = nextConsumer
+
+	return receiver, nil
+}
+
+func (f *ExampleReceiverFactory) CreateLogsReceiver(
+	_ context.Context,
+	_ component.ReceiverCreateParams,
+	cfg configmodels.Receiver,
+	nextConsumer consumer.LogsConsumer,
+) (component.LogsReceiver, error) {
+	receiver := f.createReceiver(cfg)
+	receiver.LogConsumer = nextConsumer
+
+	return receiver, nil
+}
+
+// ExampleReceiverProducer allows producing traces and metrics for testing purposes.
+type ExampleReceiverProducer struct {
+	Started         bool
+	Stopped         bool
+	TraceConsumer   consumer.TracesConsumer
+	MetricsConsumer consumer.MetricsConsumer
+	LogConsumer     consumer.LogsConsumer
+}
+
+// Start tells the receiver to start its processing.
+func (erp *ExampleReceiverProducer) Start(_ context.Context, _ component.Host) error {
+	erp.Started = true
+	return nil
+}
+
+// Shutdown tells the receiver that should stop reception,
+func (erp *ExampleReceiverProducer) Shutdown(context.Context) error {
+	erp.Stopped = true
+	return nil
+}
+
+// This is the map of already created example receivers for particular configurations.
+// We maintain this map because the ReceiverFactory is asked trace and metric receivers separately
+// when it gets CreateTracesReceiver() and CreateMetricsReceiver() but they must not
+// create separate objects, they must use one Receiver object per configuration.
+var exampleReceivers = map[configmodels.Receiver]*ExampleReceiverProducer{}
